pcie/tdisp: Define Trusted IO interface
This is for passing through PCI devices to confidential VMs with
hardware-enabled PCIe trafic encryption, the feature called
"Trusted IO" ("TIO") or "Secure IO" (no acronym), and the protocol stack
includes things as IDE and TDISP.

The TDISP setup includes a step where a passed through PCI function is
bound to a Secure VM. Define an interface for the bind() hook. VFIO-PCI is
going to implement it in the following patches.

TODO: inherit it from INTERFACE_PCIE_DEVICE as TIO only exists in PCIe?

Signed-off-by: Alexey Kardashevskiy <aik@amd.com>