vfio/iommufd: "Pin" private memory
When QEMU sets up a VFIO-PCI device on the VFIO IOMMU Type1 system, it pins
the entire guest RAM. In case of a confidential VM backed with the memfd memory,
this leads to allocating the same amount of RAM as shared, and this shared memory
gets mapped in IOMMU.

The new PCIe TDISP feature allows DMA to private memory in order to enable which
the TIO-enabled IOMMUFD driver on the host accepts a KVM VM fd. Having this handle,
the IOMMUFD driver reuses the already beforehand pinned guest private memory instead
of allocating shared pages via get_user_pages(_fast)().

The new behavior is controlled by a new vfio-pci::x-tio-pvt parameter.

Signed-off-by: Alexey Kardashevskiy <aik@amd.com>