[SQUASH] Don't warn about treating unknown memory type as zero page
It will get printed much for every upstream OVMF going forward and so
isn't really that helpful for diagnosing anything special. It would be
more interesting to print such messages when SVSM is enabled, in which
case shouldn't encounter any unknown memory types, which would then be
worth warning about.

Signed-off-by: Michael Roth <michael.roth@amd.com>